Cadastro_Sementes_do_Amanha/requirements.txt
flask
flask-sqlalchemy
pandas
openpyxl        # Excel
reportlab       # PDF
python-dateutil # parsing de datas robusto